+++ data-science-lab-app/src/app/core/core.module.ts
@@ -36,6 +36,8 @@ import { RouterService, AppRouterService } from '../services/router-service';
 import { TabService, AppTabService } from '../services/tab-service';
 import { TabFactory, AppTabFactory } from '../factory/tab-factory';
 import { UserSettingService, AppUserSettingService } from '../services/user-setting-service';
+import { CreationService } from '../services/creation-service/creation.service';
+import { AppCreationService } from '../services/creation-service/app-creation.service';
 
 @NgModule({
     declarations: [
@@ -79,7 +81,8 @@ import { UserSettingService, AppUserSettingService } from '../services/user-sett
         { provide: SidebarService, useClass: AppSidebarService },
         { provide: NotificationService, useClass: AppNotificationService },
         { provide: ErrorService, useClass: AppErrorService }, 
-        { provide: UserSettingService, useClass: AppUserSettingService }
+        { provide: UserSettingService, useClass: AppUserSettingService },
+        { provide: CreationService, useClass: AppCreationService },
     ]
 })
 export class CoreModule {
