+++ data-science-lab-app/src/app/services/creation-service/creation.service.ts
@@ -1,0 +1,3 @@
+export abstract class CreationService {
+    abstract subscribe();
+}
