+++ data-science-lab-app/src/app/services/creation-service/index.ts
@@ -1,0 +1,2 @@
+export { AppCreationService } from './app-creation.service';
+export { CreationService } from './creation.service';
